+++ AirCasting/BluetoothSession/MobilePeripheralSessionManager.swift
@@ -244,36 +244,3 @@ class MobilePeripheralSessionManager {
         activeMobileSession?.peripheral == peripheral
     }
 }
-
-
-class ReconnectionController: BluetoothConnectionObserver {
-    @Injected private var mobilePeripheralManager: MobilePeripheralSessionManager
-    @Injected private var bluetoothManager: NewBluetoothManager
-    
-    init() {
-        bluetoothManager.addConnectionObserver(self)
-    }
-    
-    func didDisconnect(device: NewBluetoothManager.BluetoothDevice) {
-        guard mobilePeripheralManager.activeSessionInProgressWith(device.peripheral) else { return } // TODO: Move away from CB!
-        mobilePeripheralManager.markActiveSessionAsDisconnected(peripheral: device.peripheral)
-        
-        bluetoothManager.connect(to: device, timeout: 10) { result in
-            switch result {
-            case .success:
-                Log.info("Reconnected to a peripheral: \(device.peripheral)")
-                self.bluetoothManager.discoverCharacteristics(for: device, timeout: 10) { result in
-                    switch result {
-                    case .success:
-                        Log.info("Discovered characteristics for: \(device.peripheral)")
-                        self.mobilePeripheralManager.configureAB()
-                    case .failure(_):
-                        self.mobilePeripheralManager.moveSessionToStandaloneMode(peripheral: device.peripheral)
-                    }
-                }
-            case .failure(_):
-                self.mobilePeripheralManager.moveSessionToStandaloneMode(peripheral: device.peripheral)
-            }
-        }
-    }
-}
